+++ packages/experimental/src/global/js/package-settings.js
@@ -28,6 +28,8 @@ const defaults = {
     ExportModal: false,
     HTTPErrors: false,
     HTTPError403: false,
+    HTTPError404: false,
+    HTTPErrorOther: false,
     ImportModal: false,
     ModifiedTabs: false,
     Notifications: false,
@@ -49,17 +51,64 @@ const defaults = {
   },
 };
 
-const setAll = (flags, enabled) => {
-  for (const key in flags) {
-    flags[key] = enabled;
-  }
-};
+const component = { ...defaults.component };
+const feature = { ...defaults.feature };
+const warningMessageComponent = (property) =>
+  `IBM Cloud Cognitive (WARNING): Component "${property}" enabled via feature flags. This component has not yet completed its review process.`;
+const warningMessageFeature = (property) =>
+  `IBM Cloud Cognitive (WARNING): Feature "${property}" enabled via feature flags.`;
+const warningMessageAllComponents =
+  'IBM Cloud Cognitive (WARNING): All components enabled through use of setAllComponents. This includes components that have not yet completed their review process.';
+const warningMessageAllFeatures =
+  'IBM Cloud Cognitive (WARNING): All features enabled through use of setAllFeatures';
 
 const settings = {
+  _allComponents: false,
+  _allFeatures: false,
   // flags
   prefix: defaults.prefix,
-  component: { ...defaults.component },
-  feature: { ...defaults.feature },
+  component: new Proxy(component, {
+    set(target, property, value) {
+      if (value) {
+        console.warn(warningMessageComponent(property));
+      }
+      target[property] = value;
+      return true; // value set
+    },
+    get(target, property) {
+      if (settings._allComponents) {
+        return true;
+      }
+
+      const propValue = target[property];
+      if (typeof propValue !== 'undefined') {
+        return propValue;
+      } else {
+        return false; // unknown component
+      }
+    },
+  }),
+  feature: new Proxy(feature, {
+    set(target, property, value) {
+      if (value) {
+        console.warn(warningMessageFeature(property));
+      }
+      target[property] = value;
+      return true; // value set
+    },
+    get(target, property) {
+      if (settings._allFeatures) {
+        return true;
+      }
+
+      const propValue = target[property];
+      if (typeof propValue !== 'undefined') {
+        return propValue;
+      } else {
+        return false; // unknown feature
+      }
+    },
+  }),
   // methods
   isComponentEnabled: (component, byDefault = false) => {
     const componentName = component?.name || component;
@@ -71,11 +120,35 @@ const settings = {
     return flags.feature[feature];
   },
   setAllComponents: (enabled) => {
-    setAll(settings.component, enabled);
+    if (enabled) {
+      console.warn(warningMessageAllComponents);
+    }
+    settings._allComponents = enabled;
   },
   setAllFeatures: (enabled) => {
-    setAll(settings.feature, enabled);
+    if (enabled) {
+      console.warn(warningMessageAllFeatures);
+    }
+    settings._allFeatures = enabled;
   },
 };
 
-export default settings;
+const settingsProxy = new Proxy(settings, {
+  set(target, property, value) {
+    switch (property) {
+      case '_allComponentsEnabled':
+        target.setAllComponents(!!value);
+        return true; // value set
+
+      case '_allFeaturesEnabled':
+        target.setAllFeatures(!!value);
+        return true; // value set
+
+      default:
+        // do nothing
+        return false; // did not set anything
+    }
+  },
+});
+
+export default settingsProxy;